changed the cart functions file to medicineViewModel
