+++ data/src/main/java/com/example/data/repository/UserRepoImpl.kt
@@ -88,34 +88,4 @@ class UserRepoImpl(private val userDAO: UserDAO) {
     }
 
     suspend fun getUserData(): UserInfo = userDAO.getUserData()
-
-    suspend fun getUserCartItems(
-        userId: String
-    ): CartResponse = UserRetrofitInstance.userApi.getUsersCartItems(userId)
-
-    suspend fun addProductToCart(
-        userId: String,
-        productId: PatchProductId
-    ): PatchRequestResponse = UserRetrofitInstance.userApi.addProductToCart(userId, productId)
-
-    suspend fun removeProductFromCart(
-        userId: String,
-        productId: PatchProductId
-    ): PatchRequestResponse = UserRetrofitInstance.userApi.removeProductFromCart(userId, productId)
-
-    suspend fun incrementProductNumberInCart(
-        userId: String,
-        productId: String
-    ): PatchRequestResponse = UserRetrofitInstance.userApi.incrementProductNumberInCart(userId, productId)
-
-    suspend fun decrementProductNumberInCart(
-        userId: String,
-        productId: String
-    ): PatchRequestResponse = UserRetrofitInstance.userApi.decrementProductNumberInCart(userId, productId)
-
-    suspend fun getNumberOfItemInCart(
-        userId: String,
-        productId: String
-    ): Int = UserRetrofitInstance.userApi.getNumberOfItemInCart(userId, productId)
-
 }
